// src/model/UserModel.ts
export class UserModel {
    nodeId: string = null;
    name: string = null;
    email: string = null;
    password: string = null;
    createdOn: string = null;
    lastLogin: string = null;
    lastGraph: string = null;
    licenseExpiration: string = null;
    profilePicture: string = null;
    domainId: string = null;
    token: string = null;
    active: number = null;
}